refactor: 우승자 이름 변환기능 분리
- 기존 Winners 클래스에서 객체 생성시 수행되던 이름 추출후 형식에 맞게 변환하는 기능을 별도의 util 클래스로 구현
- Winners는 결과 값만 가지고 있는 DTO로 수정.
