+++ src/main/java/racingcar/dto/Winners.java
@@ -1,0 +1,5 @@
+package racingcar.dto;
+
+public record Winners(String name) {
+
+}
